services/orchestrator/requirements.txt
anthropic>=0.28.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.25.0
pydantic>=2.0.0
python-dotenv>=1.0.0
